correção dos tipo de número de telefone para double

+++ src/cliente.cpp
@@ -4,7 +4,7 @@
 
 Cliente::Cliente(//implementação do construtor da classe Cliente
     unsigned int codigo,
-    unsigned int telefone,
+    double telefone,
     unsigned int pontos,
     std::string nome,
     std::string endereco
